lambda_functions/requirements.txt
psycopg2-binary
sentry-sdk
